Twentythreeth commit including: update paymentByCard and paymentByCash methods of the CartServiceImpl class. Update updateById method of the CartLineItemServiceImpl class. Add OrderItem class to model package. Add OrderItemRepository implement to the repository package. Add OrderDTOResponse, OrderItemDTOResponse, OrderDTORequest and OrderItemDTORequest methods to the DTO package. Add and handle OrderDTOResponse method to the OrderDTOResponse class. Add and handle OrderItemDTOResponse method of the OrderItemDTOResponse class. Add OrderService and OrderItemService implements to the service package. Add OrderServiceImpl and OrderItemServiceImpl class to the impl package. Add and handle findById and findAllByOrderId method to the OrderItemServiceImpl. Add addByOrderId method to the OrderItemServiceImpl class. Add and function findById, myOrder, setStatusDelivery, cancel, reactivateById, deleteById and restoreById methods to the OrderServiceImpl class. Add add and updateById methods to the OrderServiceImpl class.

+++ src/main/java/com/r2s/R2Sshop/DTO/CartLineItemSimpleDTOResponse.java
@@ -6,12 +6,14 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.springframework.util.ObjectUtils;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class CartLineItemSimpleDTOResponse {
     private Integer quantity;
-    private Double totalPrice;
+    private BigDecimal totalPrice;
     /**
      * Customize the output cartLineItem information as a JSON file.
      * <p>
